pyproject.toml
[project]
name = "voice-notes"
version = "0.1.0"
description = "A self-hosted, multi-tenant voice note application with AI transcription and semantic search"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.0",
    "sqlmodel>=0.0.22",
    "alembic>=1.13.0",
    "pydantic-settings>=2.5.0",
    "python-jose[cryptography]>=3.3.0",
    "bcrypt>=4.0.0",
    "python-multipart>=0.0.9",
    "mlx-whisper>=0.4.0",
    "numpy>=2.0.0",
    "httpx>=0.27.0",
    "sqlite-vec>=0.1.6",
    "jinja2>=3.1.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
